requirements.txt
requests
pandas
sqlalchemy
psycopg[binary]
python-dotenv
web3
eth-abi
